+++ src/test/java/page/LoginPage.java
@@ -8,11 +8,9 @@ import org.openqa.selenium.support.How;
 import org.openqa.selenium.support.PageFactory;
 import test.BaseTest;
 
-public class LoginPage extends BaseTest {
-
+public class LoginPage1 extends BaseTest {
     public WebDriver pageDriver;
-
-    public LoginPage(WebDriver driver) {
+    public LoginPage1(WebDriver driver){
         pageDriver = driver;
         PageFactory.initElements(pageDriver, this);
     }
@@ -20,28 +18,27 @@ public class LoginPage extends BaseTest {
     @FindBy( how = How.XPATH, using = LoginLocator.TITLE_PAGE)
     static WebElement title;
 
-    @FindBy( how = How.XPATH, using = LoginLocator.USER_INPUT)
-    static WebElement userNameInput;
+    @FindBy( how = How.NAME, using = LoginLocator.USER_INPUT)
+    WebElement userNameInput;
 
-    @FindBy( how = How.XPATH, using = LoginLocator.PASSWORD_INPUT)
-    static WebElement passwordInput;
+    @FindBy( how = How.NAME, using = LoginLocator.PASSWORD_INPUT)
+    WebElement passwordInput;
 
-    @FindBy( how = How.XPATH, using = LoginLocator.FORGOT_PASSWORD_LINK)
+    @FindBy( how = How.CLASS_NAME, using = LoginLocator.FORGOT_PASSWORD_LINK)
     static WebElement forgotPasswordLink;
 
-    @FindBy( how = How.XPATH, using = LoginLocator.REGISTER_LINK)
+    @FindBy( how = How.CLASS_NAME, using = LoginLocator.REGISTER_LINK)
     static WebElement RegisterLink;
 
     @FindBy( how = How.XPATH, using = LoginLocator.SUBMIT_BUTTON)
     static WebElement SubmitLink;
 
-    public static boolean sendUsername(String username) {
+    public static void sendUsername(String username) {
         System.out.print("Looking for 'Username' input");
         if(userNameInput.isDisplayed()) {
             System.out.print(" : FOUND");
             userNameInput.sendKeys(username);
             System.out.println(" & STRING SENT");
-            return true;
         }
         else {
             System.out.println(" : NOT FOUND");
@@ -49,13 +46,12 @@ public class LoginPage extends BaseTest {
         }
     }
 
-    public static boolean sendPassword(String password) {
+    public static void sendPassword(String password) {
         System.out.print("Looking for 'Password' input");
         if(passwordInput.isDisplayed()) {
             System.out.print(" : FOUND");
             passwordInput.sendKeys(password);
             System.out.println(" & STRING SENT");
-            return true;
         }
         else {
             System.out.println(" : NOT FOUND");
@@ -63,42 +59,43 @@ public class LoginPage extends BaseTest {
         }
     }
 
-    public static boolean clickOnRegisterLink() {
-        System.out.print("Looking for register link");
-        if(RegisterLink.isDisplayed()) {
-            System.out.println(" : DONE");
-            RegisterLink.click();
-            return true;
-        }
-        else {
-            System.out.println(" : NOT FOUND");
-            throw new RuntimeException();
-        }
-    }
-
-    public static boolean clickOnForgotLink() {
-        System.out.print("Looking for forgot password link");
-        if(forgotPasswordLink.isDisplayed()) {
-            System.out.println(" : DONE");
-            forgotPasswordLink.click();
-            return true;
-        }
-        else {
-            System.out.println(" : NOT FOUND");
-            throw new RuntimeException();
-        }
-    }
-
-    public static boolean clickSubmitButton() {
-        System.out.print("Looking for submit button");
+    public boolean clickOnSubmitButton() {
+        System.out.print("Looking for 'Submit' button");
         if(SubmitLink.isDisplayed()) {
-            System.out.println(" : DONE");
+            System.out.print(" : FOUND");
             SubmitLink.click();
+
+            System.out.print(" & CLICKED");
             return true;
         }
         else {
             System.out.println(" : NOT FOUND");
             throw new RuntimeException();
         }
+    }
+
+    public boolean clickOnForgotPasswordButton() {
+        System.out.print("Looking for 'Forgot Password' button");
+        if(forgotPasswordLink.isDisplayed()) {
+            System.out.print(" : FOUND");
+            forgotPasswordLink.click();
+            System.out.print(" & CLICKED");
+        }
+        else {
+            System.out.println(" : NOT FOUND");
+            throw new RuntimeException();
+        }
+    }
+
+    public static void clickOnRegisterButton() {
+        System.out.print("Looking for 'Register' button");
+        if(RegisterLink.isDisplayed()) {
+            System.out.print(" : FOUND");
+            RegisterLink.click();
+            System.out.print(" & CLICKED");
+        }
+        else {
+            System.out.println(" : NOT FOUND");
+            throw new RuntimeException();
     }
 }
